requirements.txt
# requirements.txt

# Data Collection
entsoe-py>=0.6.0          # ENTSO-E API client
requests>=2.28.0          # HTTP requests
python-dotenv>=1.0.0      # Environment variables

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0           # Parquet support

# Feature Engineering
holidays>=0.40            # Dutch holidays
astral>=3.2               # Sunrise/sunset

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0           # Alternative model
imbalanced-learn>=0.11.0  # SMOTE
optuna>=3.4.0             # Hyperparameter tuning

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
calplot>=0.1.5            # Calendar heatmaps

# Dashboard
streamlit>=1.28.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0             # Code formatting
flake8>=6.1.0             # Linting